ENG-2631, ENG-2632: Update @Statistics result table to include new HSIds.
- Change the type of the siteID column from integer to long in the @Statistics
  result table.

- The value of the siteID column is now HSId.

+++ src/frontend/org/voltdb/VoltSystemProcedure.java
@@ -39,7 +39,7 @@ import org.voltdb.messaging.FragmentTaskMessage;
 public abstract class VoltSystemProcedure extends VoltProcedure {
 
     /** Standard column type for host/partition/site id columns */
-    public static final VoltType CTYPE_ID = VoltType.INTEGER;
+    public static final VoltType CTYPE_ID = VoltType.BIGINT;
 
     /** Standard column name for a host id column */
     public static final String CNAME_HOST_ID = "HOST_ID";
